Refactor order of operations when new runs are submitted so that ninja points are averaged at the right time
- Extract averageNewRunData method into separate file for use in other modules
- Remove old unused prisma invocation
- Average NP after calculating np for the new run
